+++ R/predict-trend.R
@@ -65,6 +65,41 @@ bb_predict_recruitment_trend <- function(recruitment,
   coef
 }
 
+#' Predict Calf-Cow Ratio Trend
+#'
+#' Predict calves per adult female by year as trend line.
+#' Recruitment fit object provided must be created with `year_trend = TRUE`.
+#'
+#' @inheritParams params
+#' @return A tibble of the predicted estimates.
+#' @export
+#' @family analysis
+bb_predict_calf_cow_ratio_trend <- function(recruitment,
+                                         conf_level = 0.95,
+                                         estimate = median,
+                                         sig_fig = 5) {
+  chkor_vld(.vld_fit(recruitment), .vld_fit_ml(recruitment))
+  chk_s3_class(recruitment, "bboufit_recruitment")
+  .chk_year_trend(recruitment)
+  chk_range(conf_level)
+  chk_function(estimate)
+  chk_whole_number(sig_fig)
+  
+  predicted <- predict_trend(
+    fit = recruitment,
+    derived_expr = derived_expr_recruitment_trend()
+  )
+
+  coef <- predict_coef(
+    samples = predicted$samples,
+    new_data = predicted$data,
+    conf_level = conf_level,
+    estimate = estimate,
+    sig_fig = sig_fig
+  )
+  coef
+}
+
 #' Predict Survival Trend
 #'
 #' Predict survival by year as trend line.